+++ fish_speech/utils/schema.py
