+++ tools/api_client.py
@@ -9,7 +9,7 @@ from pydub import AudioSegment
 from pydub.playback import play
 
 from fish_speech.utils.file import audio_to_bytes, read_ref_text
-from tools.schema import ServeReferenceAudio, ServeTTSRequest
+from fish_speech.utils.schema import ServeReferenceAudio, ServeTTSRequest
 
 
 def parse_args():
