+++ tools/run_webui.py
@@ -10,7 +10,7 @@ pyrootutils.setup_root(__file__, indicator=".project-root", pythonpath=True)
 
 from fish_speech.inference_engine import TTSInferenceEngine
 from tools.llama.generate import launch_thread_safe_queue
-from tools.schema import ServeTTSRequest
+from fish_speech.utils.schema import ServeTTSRequest
 from tools.vqgan.inference import load_model as load_decoder_model
 from tools.webui import build_app
 from tools.webui.inference import get_inference_wrapper
